changed the location for components folder. Note: Components folder should remain outside th client folder while all website pages should remain inside client folder
