--- app/(client)/components/theme-provider.jsx
@@ -1,13 +1,0 @@
-'use client';
-
-import { ThemeProvider as NextThemesProvider } from "next-themes"
-
-const ThemeProvider = ({ children, ...props }) => {
-  return (
-    <NextThemesProvider {...props}>
-        {children}
-    </NextThemesProvider>
-  )
-}
-
-export default ThemeProvider
