+++ app/components/Navbar.jsx
@@ -2,7 +2,7 @@
 
 import Image from 'next/image'
 import React from 'react'
-import logo from "../../../public/images/logoTa24.jpeg"
+import logo from "../../public/images/logoTa24.jpeg"
 import Link from 'next/link'
 import { usePathname } from 'next/navigation'
