+++ app/components/theme-provider.jsx
@@ -1,8 +1,13 @@
-"use client";
+'use client';
 
-import React from "react";
-import { ThemeProvider as NextThemesProvider } from "next-themes";
+import { ThemeProvider as NextThemesProvider } from "next-themes"
 
-export function ThemeProvider({ children, ...props }) {
-  return <NextThemesProvider {...props}>{children}</NextThemesProvider>;
+const ThemeProvider = ({ children, ...props }) => {
+  return (
+    <NextThemesProvider {...props}>
+        {children}
+    </NextThemesProvider>
+  )
 }
+
+export default ThemeProvider
